+++ Abot2.Tests.Unit/Abot2.Tests.Unit.csproj
@@ -18,7 +18,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Core\" />
     <Folder Include="Crawler\" />
     <Folder Include="Util\" />
     <Folder Include="Poco\" />
